Create README.md
Hey there 😎 Welcome to my little corner of the internet. Just a dude vibin’ with some code, experiments, and random ideas. Break it, fix it, learn something, or just scroll through and enjoy the chaos 🐐🔥. No rules, just code and vibes.

If you are seeing this you are a goat🔥🐐